+++ src/lib/barrels.ts
@@ -1,9 +1,10 @@
 import { PipesModule } from './pipes/pipes.module';
-import { StBreadcrumbsModule }  from './st-breadcrumbs/st-breadcrumbs.module';
+import { StBreadcrumbsModule } from './st-breadcrumbs/st-breadcrumbs.module';
 import { StButtonModule } from './st-button/st-button.module';
 import { StDropdownModule } from './st-dropdown/st-dropdown.module';
 import { StDropdownMenuModule } from './st-dropdown-menu/st-dropdown-menu.module';
 import { StFooterModule } from './st-footer/st-footer.module';
+import { StFormLabelModule } from './utils/egeo-form/st-form-label/st-form-label.module';
 import { StHeaderModule } from './st-header/st-header.module';
 import { StHorizontalTabsModule } from './st-horizontal-tabs/st-horizontal-tabs.module';
 import { StInfoBoxModule } from './st-info-box/st-info-box.module';
@@ -18,10 +19,13 @@ import { StSearchModule } from './st-search/st-search.module';
 import { StSpinnerModule } from './st-spinner/st-spinner.module';
 import { StTabBoxModule } from './st-tab-box/st-tab-box.module';
 import { StTableModule } from './st-table/st-table.module';
+import { StTextareaModule } from './st-textarea/st-textarea.module';
 import { StToggleButtonsModule } from './st-toggle-buttons/st-toggle-buttons.module';
 import { StTooltipModule } from './st-tooltip/st-tooltip.module';
 import { StTwoListSelectionModule } from './st-two-list-selection/st-two-list-selection.module';
 import { StVerticalTabsModule } from './st-vertical-tabs/st-vertical-tabs.module';
+import { StSwitchModule } from './st-switch/st-switch.module';
+import { StCheckboxModule } from './st-checkbox/st-checkbox.module';
 
 export const DECLARATIONS: any[] = [
    PipesModule,
@@ -30,6 +34,7 @@ export const DECLARATIONS: any[] = [
    StDropdownModule,
    StDropdownMenuModule,
    StFooterModule,
+   StFormLabelModule,
    StHeaderModule,
    StHorizontalTabsModule,
    StInfoBoxModule,
@@ -46,6 +51,9 @@ export const DECLARATIONS: any[] = [
    StVerticalTabsModule,
    StPaginationModule,
    StModalModule,
+   StRadioModule,
+   StSwitchModule,
    StTableModule,
-   StRadioModule
+   StTextareaModule,
+   StCheckboxModule
 ];