+++ src/lib/st-table/index.ts
@@ -1,2 +1,0 @@
-export { StTableModule } from './st-table.module';
-export { Order, ORDER_TYPE } from './shared/order';